refactor: typescript migration without types
