+++ tsconfig.json
@@ -1,0 +1,6 @@
+{
+	"compilerOptions": {
+		
+	},
+	"include": ["src"]
+}
